+++ src/resolution.ts
@@ -1,0 +1,63 @@
+// tslint:disable-next-line:no-implicit-dependencies
+import * as ts from 'typescript'; // Imported for types alone
+
+export type ResolveModuleName = (
+  typescript: typeof ts,
+  moduleName: string,
+  containingFile: string,
+  compilerOptions: ts.CompilerOptions,
+  moduleResolutionHost: ts.ModuleResolutionHost
+) => ts.ResolvedModuleWithFailedLookupLocations;
+
+export type ResolveTypeReferenceDirective = (
+  typescript: typeof ts,
+  typeDirectiveName: string,
+  containingFile: string,
+  compilerOptions: ts.CompilerOptions,
+  moduleResolutionHost: ts.ModuleResolutionHost
+) => ts.ResolvedTypeReferenceDirectiveWithFailedLookupLocations;
+
+export function makeResolutionFunctions(
+  resolveModuleName: ResolveModuleName | undefined,
+  resolveTypeReferenceDirective: ResolveTypeReferenceDirective | undefined
+) {
+  resolveModuleName =
+    resolveModuleName ||
+    ((
+      // tslint:disable-next-line:no-shadowed-variable
+      typescript,
+      moduleName,
+      containingFile,
+      // tslint:disable-next-line:no-shadowed-variable
+      compilerOptions,
+      moduleResolutionHost
+    ) => {
+      return typescript.resolveModuleName(
+        moduleName,
+        containingFile,
+        compilerOptions,
+        moduleResolutionHost
+      );
+    });
+
+  resolveTypeReferenceDirective =
+    resolveTypeReferenceDirective ||
+    ((
+      // tslint:disable-next-line:no-shadowed-variable
+      typescript,
+      typeDirectiveName,
+      containingFile,
+      // tslint:disable-next-line:no-shadowed-variable
+      compilerOptions,
+      moduleResolutionHost
+    ) => {
+      return typescript.resolveTypeReferenceDirective(
+        typeDirectiveName,
+        containingFile,
+        compilerOptions,
+        moduleResolutionHost
+      );
+    });
+
+  return { resolveModuleName, resolveTypeReferenceDirective };
+}