+++ src/service.ts
@@ -33,6 +33,15 @@ export const createNormalizedMessageFromDiagnostic = makeCreateNormalizedMessage
 export const createNormalizedMessageFromRuleFailure = makeCreateNormalizedMessageFromRuleFailure();
 export const createNormalizedMessageFromInternalError = makeCreateNormalizedMessageFromInternalError();
 
+const resolveModuleName = process.env.RESOLVE_MODULE_NAME
+  ? require(process.env.RESOLVE_MODULE_NAME!).resolveModuleName
+  : undefined;
+const resolveTypeReferenceDirective = process.env
+  .RESOLVE_TYPE_REFERENCE_DIRECTIVE
+  ? require(process.env.RESOLVE_TYPE_REFERENCE_DIRECTIVE!)
+      .resolveTypeReferenceDirective
+  : undefined;
+
 const checker: IncrementalCheckerInterface =
   process.env.USE_INCREMENTAL_API === 'true'
     ? new ApiIncrementalChecker(
@@ -44,7 +53,9 @@ const checker: IncrementalCheckerInterface =
         process.env.CONTEXT!,
         process.env.TSLINT === 'true' ? true : process.env.TSLINT! || false,
         process.env.TSLINTAUTOFIX === 'true',
-        process.env.CHECK_SYNTACTIC_ERRORS === 'true'
+        process.env.CHECK_SYNTACTIC_ERRORS === 'true',
+        resolveModuleName,
+        resolveTypeReferenceDirective
       )
     : new IncrementalChecker(
         typescript,
@@ -59,7 +70,9 @@ const checker: IncrementalCheckerInterface =
         parseInt(process.env.WORK_NUMBER!, 10) || 0,
         parseInt(process.env.WORK_DIVISION!, 10) || 1,
         process.env.CHECK_SYNTACTIC_ERRORS === 'true',
-        process.env.VUE === 'true'
+        process.env.VUE === 'true',
+        resolveModuleName,
+        resolveTypeReferenceDirective
       );
 
 async function run(cancellationToken: CancellationToken) {